99.89% of gm_80163374

+++ src/melee/gm/gmmain_lib.c
@@ -457,7 +457,7 @@ void gmMainLib_8015D418(u8 arg0, int stocks)
     base[arg0].x7C.b13_to_15 = stocks;
 }
 
-u32* gmMainLib_8015D438(u8 arg0)
+u32* gmMainLib_8015D438(u32 arg0)
 {
     struct FighterData* base = &gmMainLib_804D3EE0->thing.x1F2C[0];
     return &base[arg0].x7C.x94;
@@ -468,7 +468,7 @@ s32* gmMainLib_8015D450(u8 arg0)
     return &gmMainLib_8015EDBC()->x4C[arg0 & 255];
 }
 
-bool gmMainLib_8015D48C(u8 arg0)
+bool gmMainLib_8015D48C(u32 arg0)
 {
     struct FighterData* base = &gmMainLib_804D3EE0->thing.x1F2C[0];
     return base[arg0].x7C.b0;
